# packages/api-types/openapi.json
{
  "components": {
    "schemas": {
      "AckResponse": {
        "properties": {
          "status": {
            "examples": [
              "acked"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "AckResponse",
        "type": "object"
      },
...
      "EmailTaskPayload": {
        "description": "Corpo da tarefa enfileirada no Cloud Tasks e entregue ao endpoint interno.",
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "to": {
            "format": "email",
            "title": "To",
            "type": "string"
          }
        },
        "required": [
          "to",
          "subject",
          "body",
          "event_id"
        ],
        "title": "EmailTaskPayload",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SETES.DOCS API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/tasks/email": {
      "post": {
        "description": "Processa uma tarefa de e-mail entregue pelo Cloud Tasks.\n\nContrato de falha (US 5.2 Cen.3): em QUALQUER falha de entrega, registra em\nlog e retorna 200 (ACK). A fila usa maxAttempts=1, então não há redespacho.\nNunca propaga exceção — isso faria o Cloud Tasks tratar como retryable.",
        "operationId": "process_email_task_internal_tasks_email_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailTaskPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AckResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Process Email Task",
        "tags": [
          "internal"
        ]
      }
    },
...
  }
}